// src/data/models/anime/AnimeType.ts
/**
 * Тип аниме
 */
export enum AnimeType {

    /**
     * ТВ-сериал
     */
    TV = "tv",

    /**
     * Полнометражный фильм
     */
    MOVIE = "movie",

    /**
     * Специальная серия (спэшл)
     */
    SPECIAL = "special",

    /**
     * Музыкальное видео
     */
    MUSIC = "music",

    /**
     * Original Video Animation
     */
    OVA = "ova",

    /**
     * Original Net Animation
     */
    ONA = "ona",

    /**
     * 13-серийное аниме
     */
    TV_13 = "tv_13",

    /**
     * 24-серийное аниме
     */
    TV_24 = "tv_24",

    /**
     * 48-серийное аниме
     */
    TV_48 = "tv_48",

    /**
     * Пока нет типа
     */
    NONE = "none",

    /**
     * Неизвестный тип (если придёт что-то не из списка)
     */
    UNKNOWN = ""
}