// src/data/models/common/AgeRatingType.ts
/**
 * Параметры возрастного рейтинга
 */
export enum AgeRatingType {

    /**
     * нет возрастного ограничения
     */
    NONE = "none",

    /**
     * все возрасты
     */
    G = "g",

    /**
     * для детей
     */
    PG = "pg",

    /**
     * для детей от 13 лет и старше
     */
    PG_13 = "pg_13",

    /**
     * для лиц старше 17 лет (насилие и ненормативная лексика)
     */
    R = "r",

    /**
     * для взрослых - Лёгкая нагота (может также содержать насилие и ненормативную лексику)
     */
    R_PLUS = "r_plus",

    /**
     * для взрослых - Лёгкая нагота (может также содержать насилие и ненормативную лексику)
     */
    R_PLUS_TWO = "r+",

    /**
     * для взрослых - Хентай (экстремальный сексуальный контент/нагота)
     */
    RX = "rx",

    /**
     * неизвестный тип (если пришло что-то не из списка)
     */
    UNKNOWN = ""
}