// src/data/models/manga/MangaType.ts
/**
 * Тип манги
 */
export enum MangaType {

    /**
     * Японская манга
     */
    MANGA = "manga",

    /**
     * Корейская манхва
     */
    MANHWA = "manhwa",

    /**
     * Китайская маньхуа
     */
    MANHUA = "manhua",

    /**
     * Новелла
     */
    LIGHT_NOVEL = "light_novel",

    /**
     * Новелла
     */
    NOVEL = "novel",

    /**
     * Ваншот
     */
    ONE_SHOT = "one_shot",

    /**
     * Додзинси
     */
    DOUJIN = "doujin",

    /**
     * Неизвестный тип (если придёт что-то не из списка)
     */
    UNKNOWN = ""
}